+++ backend/billing/billing-api/src/main/resources/application.properties
@@ -34,4 +34,4 @@ billing.stripe.webhook_secret=${STRIPE_WEBHOOK_SECRET:dev_secret}
 billing.stripe.api_key=${STRIPE_API_KEY:sk_test_51HRYgqI1ysvdCIIx21zf8xpaQ2p1M6DCaoEd4UD1c2Yr2p0sCw7PHVkV13WWFwjYqg8blbLk0llVbmpkeADxmELt00OIkvh5bT}
 
 ## Service discovery
-auth-api/mp-rest/url=${AUTH_SERVICE_BASE_URL:http://localhost:8080}
+auth-api/mp-rest/url=${AUTH_API_BASE_URL:http://localhost:8080}
