+++ backend/session/session-api/k8/base/deployment.yaml
@@ -33,7 +33,7 @@ spec:
               value: 'true'
             - name: KAFKA_BOOTSTRAP_SERVERS
               value: PLAINTEXT://kafka:9092
-            - name: AUTH_SERVICE_BASE_URL
+            - name: AUTH_API_BASE_URL
               value: http://auth-api
             - name: POSTGRES_USER
               valueFrom:
