+++ backend/session/session-api/src/main/resources/application.properties
@@ -25,7 +25,7 @@ mp.messaging.incoming.events.connector=smallrye-kafka
 mp.messaging.incoming.events.value.deserializer=com.meemaw.events.model.incoming.serialization.UserEventDeserializer
 
 ## Service discovery
-auth-api/mp-rest/url=${AUTH_SERVICE_BASE_URL:http://localhost:8080}
+auth-api/mp-rest/url=${AUTH_API_BASE_URL:http://localhost:8080}
 
 ## Location services
 location-lookup-resource/mp-rest/url=${LOCATION_LOOKUP_SERVICE_BASE_URL:http://api.ipstack.com}
